service까지 작업물

+++ src/main/java/com/moocafe/project/dao/StoreOrderDao.java
@@ -18,7 +18,7 @@ public class StoreOrderDao {
         return storeOrderRepository.save(storeOrder);
     }
 
-    public Optional<StoreOrder> findById(long id) {
+    public Optional<StoreOrder> findById(Integer id) {
         return storeOrderRepository.findById(id);
     }
